Correctly handle Prefix PathType
By creating two separate behaviors per path. Kubernetes Ingress
expects the following behavior given a path "/foo" or "/foo/":

  - Request: "/foo"     -> Matches
  - Request: "/foo/bar" -> Matches
  - Request: "/foobar"  -> No match

So instead of simply appending "*" to the path (which would match
the last case from above) we create two separate behaviors with
path patterns:

  - /foo
  - /foo/*

A special use case is "/", which results in a single behavior with
"/*" as path pattern.

https://kubernetes.io/docs/concepts/services-networking/ingress/#examples

Signed-off-by: Lucas Caparelli <lucas.caparelli@gympass.com>