Don't use TXT prefix and check all TXT values for record ownership
Since prefixes didn't work very well for alternate domains that were
also the root of their DNS zone. For example, "foo.com" from the
"foo.com" zone would end up with a TXT record "prefix-foo.com", which
is not valid for the "foo.com" zone, but is valid for the "com" zone.

Now we iterate over each value of a TXT record to determine ownership.

Signed-off-by: Lucas Caparelli <lucas.caparelli@gympass.com>