Remove CDNStatus state dependency for syncing Distributions
Depending on the existence of CDNStatus to determine it’s an update, not
a creation, makes self-healing impossible. Whenever it gets deleted, the
controller identifies an undue creation operation.

This change ensures the controller is able to reconstruct  Distribution
state through observation and not depend on CDNStatus, including for
discovering Ingresses that belong to it.

Signed-off-by: Lucas Caparelli <lucas.caparelli@gympass.com>