fix(dist-repo): create Distribution before syncing OACs, then update it
So that the Distribution is correctly configured right after Create is
called. The way it was before, OACs would be created, but would not be
linked in each origin that required them.

PostCreationOperationsFunc was added to testability, but in production
DistRepository.Sync will always be called after creation, ensuring all
is configured as it should be.

Calling Sync is a bit overkill, but simpler than a tailored post-create
function just for OACs. It's also already covered by unit tests of its
own, which is nice.

Signed-off-by: Lucas Caparelli <lucas.caparelli@gympass.com>