fix(DistributionBuilder): make duplicate Origins impossible
By storing incoming Origins as a map. This is so that when we build the
desired Distribution while iterating over all Ingresses we don't end up
duplicating Origins in it.

This scenario is possible with user-origins, where a single Origin can
be referenced by multiple Ingresses (as each user-origin is interpreted
as an Ingress).

Without this, user-origins referencing the same origin would end up
duplicated and caused duplicated behaviors/paths to be present in the
desired state.

Signed-off-by: Lucas Caparelli <lucas.caparelli@gympass.com>